refactor: extract file handling functions into useFileHandling.js
